8.1. Оптимизируй это (Часть 1) Доработки + Рабочая без редюссеров и Memo
