+++ project/src/types/state.ts
@@ -3,19 +3,12 @@ import {City} from './city';
 import {AuthorizationStatus} from '../constants';
 import {ReviewType} from './reviewType';
 
-type State = {
+export interface CounterState {
   currentCity: City,
   offers: OfferType[],
-  fetchedOffers: OfferType[],
-  currentOffer: OfferType | null,
-  authorizationStatus: AuthorizationStatus,
-  reviews: ReviewType[],
-  offerStarRating: number,
-  commentValueText: string,
-  currentSortType: string,
-  isSortingListOpen: boolean,
-  isDataLoaded: boolean,
-  observingOffer: OfferType | undefined,
-};
-
-export type {State}
+  pickedOffers: OfferType[],
+  authorizationStatus: string,
+  mapHoveredOffer: OfferType | null,
+  offerPageData: OfferType | undefined,
+  currentOfferComments: ReviewType[],
+}
